requirements.txt
huggingface_hub>=0.0.8
transformers>=4.6.1
datasets>=2.10.1
requests>=2.28.2
pandas>=1.1.5
numpy>=1.19.5
tqdm>=4.62.0
chardet>=4.0.0
python-magic>=0.4.27